requirements.txt
pyspark==3.5.1
psycopg2-binary==2.9.9
mysql-connector-python==9.0.0
pandas